backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for Conjectural Assist with LangGraph agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "google-genai>=1.62.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "supabase>=2.27.3",
    "uvicorn[standard]>=0.40.0",
    # LangGraph Agent dependencies
    "langgraph>=1.0.8",
    "langgraph-sdk>=0.3.4",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.7",
    "openai>=1.0.0",
    "numpy>=1.26.0",
    "langchain-google-genai>=2.1.0",
    "copilotkit==0.1.78",
    # necessary for path graph cross plataform (win/linux)
    "langgraph-cli[inmem]>=0.4.14",
    # NLP dependencies
    "spacy>=3.7",
    "scikit-learn>=1.4",
    "textacy>=0.13",
    # Knowledge graph
    "networkx>=3.2",
    # spaCy models (direct URLs for CI/CD compatibility)
    "pt-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.8.0/pt_core_news_md-3.8.0-py3-none-any.whl",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
]
